requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg==3.1.18
python-dotenv==1.0.0
drf-yasg==1.21.7
pytest-django==4.7.0
black==23.11.0
gunicorn==21.2.0
django-filter==23.5
